+++ slime/mtl/Alloc.h
@@ -60,9 +60,9 @@ namespace SLIME {
 
         uint32_t wasted() const { return wasted_; }
 
-        Ref alloc(int size);
+        Ref alloc(long size);
 
-        void free(int size) { wasted_ += size; }
+        void free(long size) { wasted_ += size; }
 
         // Deref, Load Effective Address (LEA), Inverse of LEA (AEL):
         T &operator[](Ref r) {
@@ -129,7 +129,7 @@ namespace SLIME {
 
     template<class T>
     typename RegionAllocator<T>::Ref
-    RegionAllocator<T>::alloc(int size) {
+    RegionAllocator<T>::alloc(long size) {
         // printf("ALLOC called (this = %p, size = %d)\n", this, size); fflush(stdout);
         assert(size > 0);
         capacity(sz + size);
